Refactory (#2)
* delete unnecessary classes and refactor service and controller

* refactor

+++ src/main/java/ua/august/todoapp/repositories/TaskRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface TaskRepository extends JpaRepository<Task, Integer> {
-    List<Task> findByOwnerId(Long id);
+    List<Task> findByOwnerId(Integer id);
 }
